# discuss: fdatasync over fsync in WAL