@ahrtr 2024-10-23
> I opened this PR to get an understanding into why `fsync` is used in the WAL over `fdatasync`

- Firstly, wal doesn't use `fsync` for all places. It just uses `fsync` when it creates, close/cut or repair a WAL file. In most cases, it uses `fdatasync`. The impact on performance should be minimal. (Note I'd like to see [rw-heatmaps](https://github.com/etcd-io/etcd/tree/main/tools/rw-heatmaps) result if you have any concern on performance).  

https://github.com/etcd-io/etcd/blob/867bd2689f5542afffb3e34de5dcb541150fca29/server/storage/wal/wal.go#L979

https://github.com/etcd-io/etcd/blob/867bd2689f5542afffb3e34de5dcb541150fca29/server/storage/wal/wal.go#L837
- The comment of `Fdatasync` already explains the reason https://github.com/etcd-io/etcd/blob/867bd2689f5542afffb3e34de5dcb541150fca29/client/pkg/fileutil/sync_linux.go#L29-L31

Please raise [discussion](https://github.com/etcd-io/etcd/discussions) for any questions. Thanks